chore: rename android module name
