+++ android/src/main/java/com/adyenreactnativesdk/AdyenDropInService.java
@@ -4,7 +4,7 @@
  * This file is open source and available under the MIT license. See the LICENSE file for more info.
  */
 
-package com.adyenreact;
+package com.adyenreactnativesdk;
 
 import android.util.Log;
