+++ android/src/main/java/com/adyenreactnativesdk/ReactNativeError.java
@@ -1,4 +1,10 @@
-package com.adyenreact;
+/*
+ * Copyright (c) 2022 Adyen N.V.
+ *
+ * This file is open source and available under the MIT license. See the LICENSE file for more info.
+ */
+
+package com.adyenreactnativesdk;
 
 import com.facebook.react.bridge.ReadableMap;
 import com.facebook.react.bridge.WritableMap;
