+++ example/android/app/src/main/java/com/adyenexample/MainApplication.java
@@ -3,7 +3,7 @@ package com.adyenexample;
 import android.app.Application;
 import android.content.Context;
 
-import com.adyenreact.AdyenPaymentPackage;
+import com.adyenreactnativesdk.AdyenPaymentPackage;
 import com.facebook.react.PackageList;
 import com.facebook.react.ReactApplication;
 import com.facebook.react.ReactInstanceManager;
